Fix ERC721 Feature detection
