+++ docs/feature_snippets.json
@@ -298,6 +298,51 @@
       "javascript": "https://docs.thirdweb.com/typescript/sdk.Erc1155Mintable"
     }
   },
+  "ERC1155SignatureMintable": {
+    "name": "Erc1155SignatureMintable",
+    "summary": "Enables generating dynamic ERC1155 NFTs with rules and an associated signature, which can then be minted by anyone securely\n\n\n",
+    "remarks": null,
+    "examples": {},
+    "methods": [
+      {
+        "name": "generate",
+        "summary": "Generate a signature that can be used to mint an NFT dynamically.\n\n",
+        "remarks": "\n\nTakes in an NFT and some information about how it can be minted, uploads the metadata and signs it with your private key. The generated signature can then be used to mint an NFT using the exact payload and signature generated.\n\n",
+        "examples": {
+          "javascript": "const nftMetadata = {\n  name: \"Cool NFT #1\",\n  description: \"This is a cool NFT\",\n  image: fs.readFileSync(\"path/to/image.png\"), // This can be an image url or file\n};\n\nconst startTime = new Date();\nconst endTime = new Date(Date.now() + 60 * 60 * 24 * 1000);\nconst payload = {\n  metadata: nftMetadata, // The NFT to mint\n  to: {{wallet_address}}, // Who will receive the NFT (or AddressZero for anyone)\n  quantity: 2, // the quantity of NFTs to mint\n  price: 0.5, // the price per NFT\n  currencyAddress: NATIVE_TOKEN_ADDRESS, // the currency to pay with\n  mintStartTime: startTime, // can mint anytime from now\n  mintEndTime: endTime, // to 24h from now\n  royaltyRecipient: \"0x...\", // custom royalty recipient for this NFT\n  royaltyBps: 100, // custom royalty fees for this NFT (in bps)\n  primarySaleRecipient: \"0x...\", // custom sale recipient for this NFT\n};\n\nconst signedPayload = contract.signature.generate(payload);\n// now anyone can use these to mint the NFT using `contract.signature.mint(signedPayload)`"
+        },
+        "reference": {
+          "javascript": "https://docs.thirdweb.com/typescript/sdk.Erc1155SignatureMintable.generate"
+        }
+      },
+      {
+        "name": "generateFromTokenId",
+        "summary": "Generate a signature that can be used to mint additionaly supply to an existing NFT.\n\n",
+        "remarks": "\n\nTakes in a payload with the token ID of an existing NFT, and signs it with your private key. The generated signature can then be used to mint additional supply to the NFT using the exact payload and signature generated.\n\n",
+        "examples": {
+          "javascript": "const nftMetadata = {\n  name: \"Cool NFT #1\",\n  description: \"This is a cool NFT\",\n  image: fs.readFileSync(\"path/to/image.png\"), // This can be an image url or file\n};\n\nconst startTime = new Date();\nconst endTime = new Date(Date.now() + 60 * 60 * 24 * 1000);\nconst payload = {\n  tokenId: 0, // Instead of metadata, we specificy the token ID of the NFT to mint supply to\n  to: {{wallet_address}}, // Who will receive the NFT (or AddressZero for anyone)\n  quantity: 2, // the quantity of NFTs to mint\n  price: 0.5, // the price per NFT\n  currencyAddress: NATIVE_TOKEN_ADDRESS, // the currency to pay with\n  mintStartTime: startTime, // can mint anytime from now\n  mintEndTime: endTime, // to 24h from now\n  royaltyRecipient: \"0x...\", // custom royalty recipient for this NFT\n  royaltyBps: 100, // custom royalty fees for this NFT (in bps)\n  primarySaleRecipient: \"0x...\", // custom sale recipient for this NFT\n};\n\nconst signedPayload = contract.signature.generate(payload);\n// now anyone can use these to mint the NFT using `contract.signature.mint(signedPayload)`"
+        },
+        "reference": {
+          "javascript": "https://docs.thirdweb.com/typescript/sdk.Erc1155SignatureMintable.generateFromTokenId"
+        }
+      },
+      {
+        "name": "mint",
+        "summary": "Mint a dynamically generated NFT\n\n",
+        "remarks": "\n\nMint a dynamic NFT with a previously generated signature.\n\n",
+        "examples": {
+          "javascript": "// see how to craft a payload to sign in the `generate()` documentation\nconst signedPayload = contract.signature.generate(payload);\n\n// now anyone can mint the NFT\nconst tx = contract.signature.mint(signedPayload);\nconst receipt = tx.receipt; // the mint transaction receipt\nconst mintedId = tx.id; // the id of the NFT minted"
+        },
+        "reference": {
+          "javascript": "https://docs.thirdweb.com/typescript/sdk.Erc1155SignatureMintable.mint"
+        }
+      }
+    ],
+    "properties": [],
+    "reference": {
+      "javascript": "https://docs.thirdweb.com/typescript/sdk.Erc1155SignatureMintable"
+    }
+  },
   "ERC20": {
     "name": "Erc20",
     "summary": "Standard ERC20 Token functions\n\n",
