+++ docs/sdk.erc1155signaturemintable._constructor_.md
@@ -9,14 +9,14 @@ Constructs a new instance of the `Erc1155SignatureMintable` class
 <b>Signature:</b>
 
 ```typescript
-constructor(contractWrapper: ContractWrapper<TContract>, storage: IStorage, roles?: ContractRoles<TokenERC1155, typeof NFTCollection.contractRoles[number]>);
+constructor(contractWrapper: ContractWrapper<BaseSignatureMintERC1155>, storage: IStorage, roles?: ContractRoles<TokenERC1155, typeof NFTCollection.contractRoles[number]>);
 ```
 
 ## Parameters
 
 |  Parameter | Type | Description |
 |  --- | --- | --- |
-|  contractWrapper | ContractWrapper&lt;TContract&gt; |  |
+|  contractWrapper | ContractWrapper&lt;BaseSignatureMintERC1155&gt; |  |
 |  storage | [IStorage](./sdk.istorage.md) |  |
 |  roles | [ContractRoles](./sdk.contractroles.md)<!-- -->&lt;TokenERC1155, typeof [NFTCollection.contractRoles](./sdk.nftcollection.contractroles.md)<!-- -->\[number\]&gt; | <i>(Optional)</i> |
