+++ docs/sdk.erc20.md
@@ -39,7 +39,7 @@ await contract.token.transfer(walletAddress, amount);
 |  [featureName](./sdk.erc20.featurename.md) |  | "ERC20" |  |
 |  [mint](./sdk.erc20.mint.md) |  | [Erc20Mintable](./sdk.erc20mintable.md) \| undefined | Mint tokens |
 |  [options](./sdk.erc20.options.md) |  | [SDKOptions](./sdk.sdkoptions.md) |  |
-|  [signature](./sdk.erc20.signature.md) |  | Erc20SignatureMintable \| undefined |  |
+|  [signature](./sdk.erc20.signature.md) |  | [Erc20SignatureMintable](./sdk.erc20signaturemintable.md) \| undefined |  |
 |  [storage](./sdk.erc20.storage.md) |  | [IStorage](./sdk.istorage.md) |  |
 
 ## Methods
