+++ docs/sdk.erc20signaturemintable._constructor_.md
@@ -1,0 +1,21 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc20SignatureMintable](./sdk.erc20signaturemintable.md) &gt; [(constructor)](./sdk.erc20signaturemintable._constructor_.md)
+
+## Erc20SignatureMintable.(constructor)
+
+Constructs a new instance of the `Erc20SignatureMintable` class
+
+<b>Signature:</b>
+
+```typescript
+constructor(contractWrapper: ContractWrapper<TokenERC20>, roles?: ContractRoles<TokenERC20, typeof Token.contractRoles[number]>);
+```
+
+## Parameters
+
+|  Parameter | Type | Description |
+|  --- | --- | --- |
+|  contractWrapper | ContractWrapper&lt;TokenERC20&gt; |  |
+|  roles | [ContractRoles](./sdk.contractroles.md)<!-- -->&lt;TokenERC20, typeof [Token.contractRoles](./sdk.token.contractroles.md)<!-- -->\[number\]&gt; | <i>(Optional)</i> |
+
