+++ docs/sdk.erc20signaturemintable.generate.md
@@ -1,8 +1,8 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc20SignatureMinting](./sdk.erc20signatureminting.md) &gt; [generate](./sdk.erc20signatureminting.generate.md)
+[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc20SignatureMintable](./sdk.erc20signaturemintable.md) &gt; [generate](./sdk.erc20signaturemintable.generate.md)
 
-## Erc20SignatureMinting.generate() method
+## Erc20SignatureMintable.generate() method
 
 Generate a signature that can be used to mint a certain amount of tokens
