+++ docs/sdk.erc20signaturemintable.generatebatch.md
@@ -1,8 +1,8 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc20SignatureMinting](./sdk.erc20signatureminting.md) &gt; [generateBatch](./sdk.erc20signatureminting.generatebatch.md)
+[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc20SignatureMintable](./sdk.erc20signaturemintable.md) &gt; [generateBatch](./sdk.erc20signaturemintable.generatebatch.md)
 
-## Erc20SignatureMinting.generateBatch() method
+## Erc20SignatureMintable.generateBatch() method
 
 Generate a batch of signatures that can be used to mint many token signatures.
 
@@ -26,5 +26,5 @@ an array of payloads and signatures
 
 ## Remarks
 
-See [Erc20SignatureMinting.generate()](./sdk.erc20signatureminting.generate.md)
+See [Erc20SignatureMintable.generate()](./sdk.erc20signaturemintable.generate.md)
