+++ docs/sdk.erc20signaturemintable.mint.md
@@ -1,8 +1,8 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc20SignatureMinting](./sdk.erc20signatureminting.md) &gt; [mint](./sdk.erc20signatureminting.mint.md)
+[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc20SignatureMintable](./sdk.erc20signaturemintable.md) &gt; [mint](./sdk.erc20signaturemintable.mint.md)
 
-## Erc20SignatureMinting.mint() method
+## Erc20SignatureMintable.mint() method
 
 Mint tokens from a signature
 
@@ -16,7 +16,7 @@ mint(signedPayload: SignedPayload20): Promise<TransactionResult>;
 
 |  Parameter | Type | Description |
 |  --- | --- | --- |
-|  signedPayload | [SignedPayload20](./sdk.signedpayload20.md) | the previously generated payload and signature with [Erc20SignatureMinting.generate()](./sdk.erc20signatureminting.generate.md) |
+|  signedPayload | [SignedPayload20](./sdk.signedpayload20.md) | the previously generated payload and signature with [Erc20SignatureMintable.generate()](./sdk.erc20signaturemintable.generate.md) |
 
 <b>Returns:</b>
