+++ docs/sdk.erc20signaturemintable.mintbatch.md
@@ -1,8 +1,8 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc20SignatureMinting](./sdk.erc20signatureminting.md) &gt; [mintBatch](./sdk.erc20signatureminting.mintbatch.md)
+[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc20SignatureMintable](./sdk.erc20signaturemintable.md) &gt; [mintBatch](./sdk.erc20signaturemintable.mintbatch.md)
 
-## Erc20SignatureMinting.mintBatch() method
+## Erc20SignatureMintable.mintBatch() method
 
 Mint any number of generated tokens signatures at once
