+++ docs/sdk.erc20signaturemintable.verify.md
@@ -1,8 +1,8 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc20SignatureMinting](./sdk.erc20signatureminting.md) &gt; [verify](./sdk.erc20signatureminting.verify.md)
+[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc20SignatureMintable](./sdk.erc20signaturemintable.md) &gt; [verify](./sdk.erc20signaturemintable.verify.md)
 
-## Erc20SignatureMinting.verify() method
+## Erc20SignatureMintable.verify() method
 
 Verify that a payload is correctly signed
