--- docs/sdk.erc20signatureminting._constructor_.md
@@ -1,21 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc20SignatureMinting](./sdk.erc20signatureminting.md) &gt; [(constructor)](./sdk.erc20signatureminting._constructor_.md)
-
-## Erc20SignatureMinting.(constructor)
-
-Constructs a new instance of the `Erc20SignatureMinting` class
-
-<b>Signature:</b>
-
-```typescript
-constructor(contractWrapper: ContractWrapper<TokenERC20>, roles: ContractRoles<TokenERC20, typeof Token.contractRoles[number]>);
-```
-
-## Parameters
-
-|  Parameter | Type | Description |
-|  --- | --- | --- |
-|  contractWrapper | ContractWrapper&lt;TokenERC20&gt; |  |
-|  roles | [ContractRoles](./sdk.contractroles.md)<!-- -->&lt;TokenERC20, typeof [Token.contractRoles](./sdk.token.contractroles.md)<!-- -->\[number\]&gt; |  |
-
