+++ docs/sdk.erc721.sig.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc721](./sdk.erc721.md) &gt; [sig](./sdk.erc721.sig.md)
+
+## Erc721.sig property
+
+<b>Signature:</b>
+
+```typescript
+sig: Erc721SignatureMintable | undefined;
+```
