--- docs/sdk.erc721.signature.md
@@ -1,11 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc721](./sdk.erc721.md) &gt; [signature](./sdk.erc721.signature.md)
-
-## Erc721.signature property
-
-<b>Signature:</b>
-
-```typescript
-signature: Erc721WithQuantitySignatureMintable<BaseSignatureMintERC721> | undefined;
-```
