+++ docs/sdk.erc721signaturemintable._constructor_.md
@@ -1,0 +1,22 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc721SignatureMintable](./sdk.erc721signaturemintable.md) &gt; [(constructor)](./sdk.erc721signaturemintable._constructor_.md)
+
+## Erc721SignatureMintable.(constructor)
+
+Constructs a new instance of the `Erc721SignatureMintable` class
+
+<b>Signature:</b>
+
+```typescript
+constructor(contractWrapper: ContractWrapper<TokenERC721>, storage: IStorage, roles?: ContractRoles<TokenERC721, typeof NFTCollection.contractRoles[number]>);
+```
+
+## Parameters
+
+|  Parameter | Type | Description |
+|  --- | --- | --- |
+|  contractWrapper | ContractWrapper&lt;TokenERC721&gt; |  |
+|  storage | [IStorage](./sdk.istorage.md) |  |
+|  roles | [ContractRoles](./sdk.contractroles.md)<!-- -->&lt;TokenERC721, typeof [NFTCollection.contractRoles](./sdk.nftcollection.contractroles.md)<!-- -->\[number\]&gt; | <i>(Optional)</i> |
+
