+++ docs/sdk.erc721signaturemintable.generate.md
@@ -1,8 +1,8 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc721SignatureMinting](./sdk.erc721signatureminting.md) &gt; [generate](./sdk.erc721signatureminting.generate.md)
+[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc721SignatureMintable](./sdk.erc721signaturemintable.md) &gt; [generate](./sdk.erc721signaturemintable.generate.md)
 
-## Erc721SignatureMinting.generate() method
+## Erc721SignatureMintable.generate() method
 
 Generate a signature that can be used to mint a dynamic NFT
