+++ docs/sdk.erc721signaturemintable.generatebatch.md
@@ -1,8 +1,8 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc721SignatureMinting](./sdk.erc721signatureminting.md) &gt; [generateBatch](./sdk.erc721signatureminting.generatebatch.md)
+[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc721SignatureMintable](./sdk.erc721signaturemintable.md) &gt; [generateBatch](./sdk.erc721signaturemintable.generatebatch.md)
 
-## Erc721SignatureMinting.generateBatch() method
+## Erc721SignatureMintable.generateBatch() method
 
 Genrate a batch of signatures that can be used to mint many dynamic NFTs.
 
@@ -26,5 +26,5 @@ an array of payloads and signatures
 
 ## Remarks
 
-See [Erc721SignatureMinting.generate()](./sdk.erc721signatureminting.generate.md)
+See [Erc721SignatureMintable.generate()](./sdk.erc721signaturemintable.generate.md)
