+++ docs/sdk.erc721signaturemintable.md
@@ -1,0 +1,30 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc721SignatureMintable](./sdk.erc721signaturemintable.md)
+
+## Erc721SignatureMintable class
+
+Enables generating dynamic ERC721 NFTs with rules and an associated signature, which can then be minted by anyone securely
+
+<b>Signature:</b>
+
+```typescript
+export declare class Erc721SignatureMintable 
+```
+
+## Constructors
+
+|  Constructor | Modifiers | Description |
+|  --- | --- | --- |
+|  [(constructor)(contractWrapper, storage, roles)](./sdk.erc721signaturemintable._constructor_.md) |  | Constructs a new instance of the <code>Erc721SignatureMintable</code> class |
+
+## Methods
+
+|  Method | Modifiers | Description |
+|  --- | --- | --- |
+|  [generate(mintRequest)](./sdk.erc721signaturemintable.generate.md) |  | Generate a signature that can be used to mint a dynamic NFT |
+|  [generateBatch(payloadsToSign)](./sdk.erc721signaturemintable.generatebatch.md) |  | Genrate a batch of signatures that can be used to mint many dynamic NFTs. |
+|  [mint(signedPayload)](./sdk.erc721signaturemintable.mint.md) |  | Mint a dynamically generated NFT |
+|  [mintBatch(signedPayloads)](./sdk.erc721signaturemintable.mintbatch.md) |  | Mint any number of dynamically generated NFT at once |
+|  [verify(signedPayload)](./sdk.erc721signaturemintable.verify.md) |  | Verify that a payload is correctly signed |
+
