+++ docs/sdk.erc721signaturemintable.mint.md
@@ -1,8 +1,8 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc721SignatureMinting](./sdk.erc721signatureminting.md) &gt; [mint](./sdk.erc721signatureminting.mint.md)
+[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc721SignatureMintable](./sdk.erc721signaturemintable.md) &gt; [mint](./sdk.erc721signaturemintable.mint.md)
 
-## Erc721SignatureMinting.mint() method
+## Erc721SignatureMintable.mint() method
 
 Mint a dynamically generated NFT
 
@@ -16,7 +16,7 @@ mint(signedPayload: SignedPayload721): Promise<TransactionResultWithId>;
 
 |  Parameter | Type | Description |
 |  --- | --- | --- |
-|  signedPayload | [SignedPayload721](./sdk.signedpayload721.md) | the previously generated payload and signature with [Erc721SignatureMinting.generate()](./sdk.erc721signatureminting.generate.md) |
+|  signedPayload | [SignedPayload721](./sdk.signedpayload721.md) | the previously generated payload and signature with [Erc721SignatureMintable.generate()](./sdk.erc721signaturemintable.generate.md) |
 
 <b>Returns:</b>
