+++ docs/sdk.erc721signaturemintable.mintbatch.md
@@ -1,8 +1,8 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc721SignatureMinting](./sdk.erc721signatureminting.md) &gt; [mintBatch](./sdk.erc721signatureminting.mintbatch.md)
+[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc721SignatureMintable](./sdk.erc721signaturemintable.md) &gt; [mintBatch](./sdk.erc721signaturemintable.mintbatch.md)
 
-## Erc721SignatureMinting.mintBatch() method
+## Erc721SignatureMintable.mintBatch() method
 
 Mint any number of dynamically generated NFT at once
