+++ docs/sdk.erc721signaturemintable.verify.md
@@ -1,8 +1,8 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc721SignatureMinting](./sdk.erc721signatureminting.md) &gt; [verify](./sdk.erc721signatureminting.verify.md)
+[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc721SignatureMintable](./sdk.erc721signaturemintable.md) &gt; [verify](./sdk.erc721signaturemintable.verify.md)
 
-## Erc721SignatureMinting.verify() method
+## Erc721SignatureMintable.verify() method
 
 Verify that a payload is correctly signed
