--- docs/sdk.erc721signatureminting._constructor_.md
@@ -1,22 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc721SignatureMinting](./sdk.erc721signatureminting.md) &gt; [(constructor)](./sdk.erc721signatureminting._constructor_.md)
-
-## Erc721SignatureMinting.(constructor)
-
-Constructs a new instance of the `Erc721SignatureMinting` class
-
-<b>Signature:</b>
-
-```typescript
-constructor(contractWrapper: ContractWrapper<TokenERC721>, roles: ContractRoles<TokenERC721, typeof NFTCollection.contractRoles[number]>, storage: IStorage);
-```
-
-## Parameters
-
-|  Parameter | Type | Description |
-|  --- | --- | --- |
-|  contractWrapper | ContractWrapper&lt;TokenERC721&gt; |  |
-|  roles | [ContractRoles](./sdk.contractroles.md)<!-- -->&lt;TokenERC721, typeof [NFTCollection.contractRoles](./sdk.nftcollection.contractroles.md)<!-- -->\[number\]&gt; |  |
-|  storage | [IStorage](./sdk.istorage.md) |  |
-
