--- docs/sdk.erc721signatureminting.md
@@ -1,30 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [Erc721SignatureMinting](./sdk.erc721signatureminting.md)
-
-## Erc721SignatureMinting class
-
-Enables generating dynamic ERC721 NFTs with rules and an associated signature, which can then be minted by anyone securely
-
-<b>Signature:</b>
-
-```typescript
-export declare class Erc721SignatureMinting 
-```
-
-## Constructors
-
-|  Constructor | Modifiers | Description |
-|  --- | --- | --- |
-|  [(constructor)(contractWrapper, roles, storage)](./sdk.erc721signatureminting._constructor_.md) |  | Constructs a new instance of the <code>Erc721SignatureMinting</code> class |
-
-## Methods
-
-|  Method | Modifiers | Description |
-|  --- | --- | --- |
-|  [generate(mintRequest)](./sdk.erc721signatureminting.generate.md) |  | Generate a signature that can be used to mint a dynamic NFT |
-|  [generateBatch(payloadsToSign)](./sdk.erc721signatureminting.generatebatch.md) |  | Genrate a batch of signatures that can be used to mint many dynamic NFTs. |
-|  [mint(signedPayload)](./sdk.erc721signatureminting.mint.md) |  | Mint a dynamically generated NFT |
-|  [mintBatch(signedPayloads)](./sdk.erc721signatureminting.mintbatch.md) |  | Mint any number of dynamically generated NFT at once |
-|  [verify(signedPayload)](./sdk.erc721signatureminting.verify.md) |  | Verify that a payload is correctly signed |
-
