+++ docs/sdk.nftcollection.md
@@ -44,6 +44,7 @@ const contract = sdk.getNFTCollection("{{contract_address}}");
 |  [roles](./sdk.nftcollection.roles.md) |  | [ContractRoles](./sdk.contractroles.md)<!-- -->&lt;TokenERC721, typeof [NFTCollection.contractRoles](./sdk.nftcollection.contractroles.md)<!-- -->\[number\]&gt; |  |
 |  [royalties](./sdk.nftcollection.royalties.md) |  | [ContractRoyalty](./sdk.contractroyalty.md)<!-- -->&lt;TokenERC721, typeof NFTCollection.schema&gt; | Configure royalties |
 |  [sales](./sdk.nftcollection.sales.md) |  | [ContractPrimarySale](./sdk.contractprimarysale.md)<!-- -->&lt;TokenERC721&gt; |  |
+|  [signature](./sdk.nftcollection.signature.md) |  | import("..").[Erc721SignatureMintable](./sdk.erc721signaturemintable.md) \| undefined |  |
 
 ## Methods
