+++ docs/sdk.nftcollection.signature.md
@@ -1,0 +1,11 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@thirdweb-dev/sdk](./sdk.md) &gt; [NFTCollection](./sdk.nftcollection.md) &gt; [signature](./sdk.nftcollection.signature.md)
+
+## NFTCollection.signature property
+
+<b>Signature:</b>
+
+```typescript
+signature: import("..").Erc721SignatureMintable | undefined;
+```
