+++ src/contracts/token.ts
@@ -3,7 +3,6 @@ import { TokenERC20 } from "contracts";
 import { ContractMetadata } from "../core/classes/contract-metadata";
 import { ContractRoles } from "../core/classes/contract-roles";
 import {
-  Erc20SignatureMinting,
   IStorage,
   NetworkOrSignerOrProvider,
   TransactionResult,
@@ -58,7 +57,6 @@ export class Token extends Erc20<TokenERC20> {
    * @internal
    */
   public analytics: ContractAnalytics<TokenERC20>;
-
   /**
    * @internal
    */
@@ -92,10 +90,6 @@ export class Token extends Erc20<TokenERC20> {
     this.events = new ContractEvents(this.contractWrapper);
     this.platformFees = new ContractPlatformFee(this.contractWrapper);
     this.interceptor = new ContractInterceptor(this.contractWrapper);
-    this.signature = new Erc20SignatureMinting(
-      this.contractWrapper,
-      this.roles,
-    );
   }
 
   /** ******************************
