+++ src/core/classes/erc-721.ts
@@ -29,7 +29,7 @@ import {
 import { FEATURE_NFT } from "../../constants/erc721-features";
 import { DetectableFeature } from "../interfaces/DetectableFeature";
 import { Erc721Dropable } from "./erc-721-dropable";
-import { Erc721WithQuantitySignatureMintable } from "./erc-721-with-quantity-signature-mintable";
+import { Erc721SignatureMintable } from "./erc-721-signature-mintable";
 
 /**
  * Standard ERC721 NFT functions
@@ -42,11 +42,8 @@ import { Erc721WithQuantitySignatureMintable } from "./erc-721-with-quantity-sig
  * @public
  */
 export class Erc721<
-  T extends
-    | Multiwrap
-    | SignatureDrop
-    | DropERC721
-    | TokenERC721
+  T extends Multiwrap | SignatureDrop | DropERC721 | TokenERC721 | BaseERC721 =
+    | BaseERC721
     | BaseSignatureMintERC721,
 > implements UpdateableNetwork, DetectableFeature
 {
@@ -54,7 +51,7 @@ export class Erc721<
   public query: Erc721Supply | undefined;
   public mint: Erc721Mintable | undefined;
   public drop: Erc721Dropable | undefined;
-  public signature: Erc721WithQuantitySignatureMintable | undefined;
+  public sig: Erc721SignatureMintable | undefined;
   protected contractWrapper: ContractWrapper<T>;
   protected storage: IStorage;
   protected options: SDKOptions;
@@ -78,7 +75,7 @@ export class Erc721<
     this.query = this.detectErc721Enumerable();
     this.mint = this.detectErc721Mintable();
     this.drop = this.detectErc721Dropable();
-    this.signature = this.detectErc721SignatureMintable();
+    this.sig = this.detectErc721SignatureMintable();
   }
 
   /**
@@ -295,19 +292,14 @@ export class Erc721<
     return undefined;
   }
 
-  private detectErc721SignatureMintable():
-    | Erc721WithQuantitySignatureMintable
-    | undefined {
+  private detectErc721SignatureMintable(): Erc721SignatureMintable | undefined {
     if (
-      detectContractFeature<BaseSignatureMintERC721>(
+      detectContractFeature<TokenERC721>(
         this.contractWrapper,
         "ERC721SignatureMint",
       )
     ) {
-      return new Erc721WithQuantitySignatureMintable(
-        this.contractWrapper,
-        this.storage,
-      );
+      return new Erc721SignatureMintable(this.contractWrapper, this.storage);
     }
     return undefined;
   }
