+++ src/core/classes/index.ts
@@ -13,10 +13,10 @@ export * from "./erc-20";
 export * from "./erc-20-mintable";
 export * from "./erc-20-batch-mintable";
 export * from "./erc-20-history";
-export * from "./erc-20-signature-minting";
+export * from "./erc-20-signature-mintable";
 export * from "./erc-721";
 export * from "./erc-721-dropable";
-export * from "./erc-721-signature-minting";
+export * from "./erc-721-signature-mintable";
 export * from "./erc-721-supply";
 export * from "./erc-721-enumerable";
 export * from "./erc-721-mintable";
