+++ test/signature-mint-20.test.ts
@@ -1,6 +1,6 @@
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
 import { assert, expect } from "chai";
-import { BigNumber, ethers } from "ethers";
+import { BigNumber } from "ethers";
 import { PayloadToSign20, SignedPayload20, Token } from "../src";
 import { sdk, signers } from "./before-setup";
 import { NATIVE_TOKEN_ADDRESS } from "../src/constants/currency";
